Api/requirements.txt
amqp==5.3.1
asgiref==3.8.1
babel==2.16.0
billiard==4.2.1
celery==5.4.0
channels==4.2.0
chardet==5.2.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Django==4.2.16
django-cors-headers==4.6.0
django-crispy-forms==2.3
django-debug-toolbar==4.4.6
django-extensions==3.2.3
django-filter==24.3
django-formtools==2.5.1
django-guardian==2.4.0
django-money==3.5.3
django-otp==1.5.4
django-phonenumber-field==8.0.0
django-postgres-extra==2.0.8
django-two-factor-auth==1.17.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
et_xmlfile==2.0.0
fonttools==4.55.0
inflection==0.5.1
kiwisolver==1.4.7
kombu==5.4.2
matplotlib==3.9.3
numpy==2.1.3
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.0.0
plotly==5.24.1
prompt_toolkit==3.0.48
py-moneyed==3.0
PyJWT==2.10.1
pyparsing==3.2.0
pypng==0.20220715.0
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
qrcode==7.4.2
reportlab==4.2.5
setuptools==75.6.0
six==1.16.0
sqlparse==0.5.2
tenacity==9.0.0
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
